+++ go/integTest/batch_test.go
@@ -159,7 +159,8 @@ func (suite *GlideTestSuite) TestBatchCommandArgsError() {
 	suite.runWithDefaultClients(func(client interfaces.BaseClientCommands) {
 		key := "{prefix}" + uuid.NewString()
 
-		opts := options.NewGetExOptions().SetExpiry(options.NewExpiry().SetType(constants.ExpiryType("pewpew")))
+		opts := options.NewGetExOptions().
+			SetExpiry(options.NewExpiryIn(10 * time.Second).SetType(constants.ExpiryType("pewpew")))
 		transaction := pipeline.NewClusterBatch(true).
 			Get(key).
 			GetExWithOptions(key, *opts).
@@ -479,9 +480,7 @@ func CreateStringTest(batch *pipeline.ClusterBatch, isAtomic bool, serverVer str
 	batch.Set(atomicKey1, value1)
 	testData = append(testData, CommandTestData{ExpectedResponse: "OK", TestName: "Set(atomicKey1, value1)"})
 	opts := options.NewGetExOptions().
-		SetExpiry(options.NewExpiry().
-			SetType(constants.Seconds).
-			SetCount(5))
+		SetExpiry(options.NewExpiryIn(5 * time.Second))
 	batch.GetExWithOptions(atomicKey1, *opts)
 	testData = append(testData, CommandTestData{ExpectedResponse: value1, TestName: "GetExWithOptions(atomicKey1, opts)"})